Do not compare signed and unsigned types
The casting of nread is safe as it is tested to be >0 when it is
initally assigned.  It is also being implicitly cast in the fwrite
call in the line above.

+++ lib/libalpm/dload.c
@@ -251,7 +251,7 @@ static int download_internal(const char *url, const char *localpath,
 		check_stop();
 		size_t nwritten = 0;
 		nwritten = fwrite(buffer, 1, nread, localf);
-		if((nwritten != nread) || ferror(localf)) {
+		if((nwritten != (size_t)nread) || ferror(localf)) {
 			pm_errno = PM_ERR_RETRIEVE;
 			_alpm_log(PM_LOG_ERROR, _("error writing to file '%s': %s\n"),
 					tempfile, strerror(errno));
